Controleur::achatCarte : pas encore tester

+++ Splendor/src/carte/carte.cpp
@@ -9,8 +9,7 @@ namespace Splendor{
     bool CarteDeveloppement::canBeBougth(Joueur &j) const {
         /*Pour pouvoir acheter un carte le joueur doit avoir les ressources
         et les bonus suffisant, et si il en manque il doit pouvoir compenser avec des jokers*/
-        int jetons_manquants = 0;
-
+        unsigned int jetons_manquants = 0;
         for (size_t i = 0; i<5;i++)
             jetons_manquants += max((unsigned int)0, getCouts(i)-j.getBonus(i)-j.getInventaire(i));
         return j.getInventaire(5)>=jetons_manquants;
@@ -18,7 +17,7 @@ namespace Splendor{
 
     bool CarteNoble::canBeBougth(Joueur &j) const {
         /*Pour acquérir un carte noble il suffit d'avoir les bonus requis*/
-        int bonus_manquants = 0;
+        unsigned int bonus_manquants = 0;
         for (size_t i = 0; i<5;i++)
             bonus_manquants += max((unsigned int)0, getCouts(i)-j.getBonus(i));
         return bonus_manquants==0;
@@ -27,10 +26,10 @@ namespace Splendor{
     bool CarteCite::canBeBougth(Joueur &j)const {
         /*Pour acquérir un carte cite il faut les bonus requis
         et les points de victoire (points de prestige) requis*/
-        int bonus_manquants = 0;
+        unsigned int bonus_manquants = 0;
         for (size_t i = 0; i<5;i++)
             bonus_manquants += max((unsigned int)0, getCouts(i)-j.getBonus(i));
-        return bonus_manquants==0 && j.getPDV() >= pdv_requis;
+        return bonus_manquants==0 && (int)j.getPDV() >= pdv_requis;
     }
 
     void CarteDeveloppement::afficherCarte(std::ostream& f) const {
